CreateAndSwitchToNewProfile was removed.
In profile_window, we were only overriding CreateAndSwitchToNewProfile
to pick a new random avatar instead of the default placeholder avatar.

The new implementation of new profile creation presents a new profile
UI where the user gets to pick the profile name and avatar, so our
override and the browser tests are no longer necessary.

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/d6329604950dbce6041d168f69df3564b841bf19

commit d6329604950dbce6041d168f69df3564b841bf19
Author: Jan Krcal <jkrcal@chromium.org>
Date:   Wed Nov 11 09:34:16 2020 +0000

    [People menus] Update the add action and rename to Profiles

    This CL updates the "Add profile..." action in profile menus on Mac and
    on linux (DBus). The CL makes it consistent with all other entry points
    to open the profile creation flow.

    This CL also updates the strings in profile menus to refer to "profiles"
    instead of "people". This change is guarded by a feature flag.

    Bug: 1086486, 1063856